code/basic_website/pyproject.toml
[project]
name = "basic_website"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sean Daly", email = "the_sean_c@fastmail.com"},
]
dependencies = [
    "python-fasthtml>=0.4.5",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
